Completata prima versione algoritmo di training.

+++ src/Constants.java
@@ -8,7 +8,7 @@ public class Constants {
     static int STIMULATION_LENGTH = 200; // Length in ms of the period during which the inputs are presented to the network.
     static int DELTA_TIME = 10; 
     static int PAUSE_LENGTH = 100; 
-    static int MAX_PIC_PIXELS = 1024; // The maximum number of pixels a sample image can be made of.
+    static short MAX_PIC_PIXELS = 1024; // The maximum number of pixels a sample image can be made of.
     
     /* Network related constants */
 	static final int IPTOS_THROUGHPUT = 0x08;
